Initial Upload of working V1
Install on Raspberry PI - forked and adjusted from https://www.instructables.com/Raspberry-Pi-NOAA-Weather-Satellite-Receiver/  And https://www.instructables.com/Raspberry-Pi-NOAA-and-Meteor-M-2-Receiver/ -  Kudos to  haslettj and NateDN10 for those. While there are some nits in the instructables the learning curve was massively shortened by thier work!

I have added lighthttpd with a very simple webpage that displays the recent images and schedule, and it also emails the latest image to zapier to then post on a facebook page i setup for it which can be found here:  https://bit.ly/2LFereD

The facebook page has photos of the antenna and RPI in use - look for the very first posts.